LayoutTests/fast/dom/shadow/shadow-root-attached-expected.txt

This test checks nodes are reallyed attached.

PASS internals.attached(light) is false
PASS internals.attached(light) is true
PASS internals.attached(light) is false
PASS internals.attached(light) is true
PASS successfullyParsed is true

TEST COMPLETE
LIGHT
